django_business_logic/requirements/base.txt
bcrypt==3.1.4
cffi==1.11.5
Django==2.0
django-appconf==1.0.2
django-compressor==2.2
django-cors-headers==2.4.0
django-crontab==0.7.1
django-extensions==2.1.0
django-filter==2.0.0
django-model-utils==3.1.2
djangorestframework==3.8.2
djangorestframework-serializer-extensions==0.6.0
hashids==1.2.0
Markdown==2.6.11
pycparser==2.18
python-decouple==3.1
pytz==2018.5
rcssmin==1.0.6
rjsmin==1.0.12
six==1.11.0
whitenoise==3.3.1
psycopg2
